Loop Fix
- Fixed a major bug in which songs wouldn't loop properly.
- Also fixed a bug where songs wouldn't loop properly if Time was set.
- Also updated the Time variable on SongEvents to instead use doubles as they are more accurate than floats.

+++ Runtime/WAVYSong.cs
@@ -56,9 +56,9 @@ namespace WAVYMusic
         public class Event
         {
             public string Name;
-            public float Time;
+            public double Time;
 
-            public Event(string name, float time)
+            public Event(string name, double time)
             {
                 Name = name;
                 Time = time;
